[kakao] use setCustomUserClaims to link accounts #10
Exhibit better ways to link custom oauth providers for users already
signed with built-in providers. The current best way to do so is
organized in the github issue #10 below:

https://github.com/firebase/custom-auth-samples/issues/10

Admin.auth.Auth.setCustomUserClaims(uid, customUserClaims) was added for
setting custom data for control access.

TODO: As mentioned in the above issue, clients should ensure that the
same user is previously signed in with built-in providers
(facebook or google), but this is omitted in the sample. It is each
developer's responsibility to ensure this for security.

Signed-off-by: Hara Kang <hara0115@gmail.com>